# Conflicts:
#	CHANGELOG.md

+++ udata_hydra/utils/errors.py
@@ -112,9 +112,10 @@ class IOException(ExceptionWithSentryDetails):
 
 async def handle_parse_exception(
     e: IOException | ParseException, table_name: str | None, check: Record | None
-) -> None:
+) -> Record | None:
     """Specific IO/ParseException handling. Store error in :check: if in a check context. Also cleanup :table_name: if needed."""
-    if table_name is not None:
+    if table_name is not None and (check and not check.get("parsing_table")):
+        # only deleting the table if we have not successfully completed csv_to_db
         db = await context.pool("csv")
         await db.execute(f'DROP TABLE IF EXISTS "{table_name}"')
         await db.execute(f"DELETE FROM tables_index WHERE parsing_table='{table_name}'")
@@ -124,10 +125,11 @@ async def handle_parse_exception(
         err = f"{e.step}:{str(e.__cause__)}"
         if e.sentry_event_id:
             err = f"{e.step}:sentry:{e.sentry_event_id}"
-        await Check.update(
+        check = await Check.update(
             check["id"],
             {"parsing_error": err, "parsing_finished_at": datetime.now(timezone.utc)},
         )
         log.error("Parsing error", exc_info=e)
+        return check
     else:
         raise e
